Refactored Dashboard to break out components
