+++ frontend/app/actions/patient/patient-search-action.ts
@@ -1,45 +1,42 @@
 'use server'
 
+import type { Patient } from '@api/models/Patient'
 import { getApiClient } from '@lib/api'
 import { authOptions } from '@lib/auth'
 import { getServerSession } from 'next-auth'
 
 /**
- * Searches for patients using a query string.
- * This is a server action that requires authentication.
+ * Server action for searching patients by name or ID
  *
- * @param {string} query - The search query to find patients
- * @returns {Promise<any>} The search results from the API
- * @throws {Error} If user is not authenticated with message 'Unauthorized'
+ * @param query - Search query (name or ID)
+ * @returns Array of matching patients
+ * @throws {Error} If user is not authenticated
  * @throws {Error} If the API request fails
  *
  * @example
  * try {
- *   const results = await searchPatients("John Doe");
+ *   const patients = await searchPatients("John");
  *   // Handle search results
  * } catch (error) {
  *   // Handle error
  * }
  */
-export async function searchPatients(query: string) {
+export async function searchPatients(query: string): Promise<Patient[]> {
+  const session = await getServerSession(authOptions)
+  if (!session) {
+    throw new Error('You must be logged in to search patients')
+  }
+
   try {
-    const session = await getServerSession(authOptions)
-    if (!session) {
-      throw new Error('Unauthorized')
-    }
-
     const api = await getApiClient(session)
-    const response = await api.request.request({
+    const searchParams = new URLSearchParams({ q: query })
+    const response = await api.request.request<Patient[]>({
       method: 'GET',
-      url: '/api/patients/search/',
-      query: {
-        q: query
-      }
+      url: `/api/patients/search/?${searchParams.toString()}`
     })
-
     return response
   } catch (error) {
-    console.error('Search error:', error)
+    console.error('Failed to search patients:', error)
     throw error
   }
 }
